RESNET-50
This script is designed for identifying and distinguishing lithology units using multi-dimensional satellite images, including radar and multispectral images from Sentinel 1 and 2. The input data is compressed using the PCA algorithm and provided to the algorithm in the form of three bands with the most information. The script includes mounting Google Drive, installing necessary libraries, preparing data, defining and training a ResNet50 model, and evaluating the model.